requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==7.1.2
colorama==0.4.4
Flask==1.1.4
flask-restx==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
pyrsistent==0.17.3
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
tzdata==2021.1
Werkzeug==1.0.1
